LookupTables renamed, Lookup table base class added
